+++ src/test/java/com/hotels/hcommon/hive/metastore/client/closeable/CloseableMetaStoreClientInvocationHandlerTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018 Expedia Inc.
+ * Copyright (C) 2018-2019 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,12 +15,16 @@
  */
 package com.hotels.hcommon.hive.metastore.client.closeable;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 import java.lang.reflect.Method;
 
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
+import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
 import org.apache.thrift.TApplicationException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -55,7 +59,6 @@ public class CloseableMetaStoreClientInvocationHandlerTest {
     String dbName = "db";
     String tableName = "table";
 
-
     IMetaStoreClient exceptionThrowingClient = Mockito.mock(IMetaStoreClient.class, new Answer() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
@@ -68,22 +71,66 @@ public class CloseableMetaStoreClientInvocationHandlerTest {
     verify(compatibility).getTable(eq(dbName), eq(tableName));
   }
 
-  @Test(expected = RuntimeException.class)
+  @Test(expected = NoSuchObjectException.class)
   public void dontInvokeCompatibilityWhenExceptionOtherThanTApplicationExceptionIsThrown() throws Throwable {
     Class<?> clazz = Class.forName(IMetaStoreClient.class.getName());
     Method method = clazz.getMethod("getTable", String.class, String.class);
     String dbName = "db";
     String tableName = "table";
 
-
-    IMetaStoreClient exceptionThrowingClient = Mockito.mock(IMetaStoreClient.class, new Answer() {
+    IMetaStoreClient exceptionThrowingClient = Mockito.mock(IMetaStoreClient.class, new Answer<Void>() {
       @Override
-      public Object answer(InvocationOnMock invocation) throws Throwable {
-        throw new RuntimeException();
+      public Void answer(InvocationOnMock invocation) throws Throwable {
+        throw new NoSuchObjectException();
       }
     });
 
     invocationHandler = new CloseableMetaStoreClientInvocationHandler(exceptionThrowingClient, compatibility);
     invocationHandler.invoke(null, method, new String[] { dbName, tableName });
+  }
+
+  @Test(expected = NoSuchObjectException.class)
+  public void rethrowCompatibilityClassException() throws Throwable {
+    Class<?> clazz = Class.forName(IMetaStoreClient.class.getName());
+    Method method = clazz.getMethod("getTable", String.class, String.class);
+    String dbName = "db";
+    String tableName = "table";
+
+    IMetaStoreClient exceptionThrowingClient = Mockito.mock(IMetaStoreClient.class, new Answer<Void>() {
+      @Override
+      public Void answer(InvocationOnMock invocation) throws Throwable {
+        throw new TApplicationException();
+      }
+    });
+
+    when(compatibility.getTable(dbName, tableName)).thenThrow(new NoSuchObjectException("This should be thrown"));
+
+    invocationHandler = new CloseableMetaStoreClientInvocationHandler(exceptionThrowingClient, compatibility);
+    invocationHandler.invoke(null, method, new String[] { dbName, tableName });
+  }
+
+  @Test
+  public void rethrowClientClassExceptionOnTApplicationException() throws Throwable {
+    Class<?> clazz = Class.forName(IMetaStoreClient.class.getName());
+    Method method = clazz.getMethod("getTable", String.class, String.class);
+    String dbName = "db";
+    String tableName = "table";
+
+    final TApplicationException original = new TApplicationException("original");
+    IMetaStoreClient exceptionThrowingClient = Mockito.mock(IMetaStoreClient.class, new Answer<Void>() {
+      @Override
+      public Void answer(InvocationOnMock invocation) throws Throwable {
+        throw original;
+      }
+    });
+
+    when(compatibility.getTable(dbName, tableName)).thenThrow(new TApplicationException("This should not be thrown"));
+
+    invocationHandler = new CloseableMetaStoreClientInvocationHandler(exceptionThrowingClient, compatibility);
+    try {
+      invocationHandler.invoke(null, method, new String[] { dbName, tableName });
+    } catch (TApplicationException e) {
+      assertThat(e, is(original));
+    }
   }
 }